refactor: move what the zone canvas draws into a file of its own
ZoneCanvas was one line under the 300-line limit and the editing half was about
to grow. The three functions that only draw, and decide nothing, move to
ZoneCanvas+Drawing.swift: one view, two files, the deciding in one of them.
